fix(CT1) rotate events correctly
This fix differentiatates the logic of the `rotate` command to operate
correctly on the CT1 Events.
The implementation relies on some lower-level APIs (CT1 models) to
provide an immediate fix for the command.

In the context of TEC and ECP, some thinking should be given to how we
would like to support Event updates like the ones performed by the
command, in the context of the ORM. I did not want to rush in a
solution that would stick for long and relied, for that reason on the
use of lower-level APIs.